+++ sdk/src/sharedTest/java/co/omise/android/ui/SecurityCodeTooltipDialogFragmentTest.kt
@@ -6,27 +6,19 @@ import android.os.Bundle
 import android.view.View
 import android.widget.ImageView
 import androidx.annotation.DrawableRes
-import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.testing.launchFragment
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.assertion.ViewAssertions.doesNotExist
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.RootMatchers.isDialog
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
-import androidx.test.espresso.matcher.ViewMatchers.isRoot
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import co.omise.android.R
 import co.omise.android.models.CardBrand
-import org.hamcrest.CoreMatchers.not
 import org.hamcrest.Description
 import org.hamcrest.Matcher
 import org.hamcrest.TypeSafeMatcher
-import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
-import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -90,24 +82,6 @@ class SecurityCodeTooltipDialogFragmentTest {
         onView(withId(R.id.cvv_description_text))
                 .inRoot(isDialog())
                 .check(matches(withText(R.string.cvv_tooltip_4_digits)))
-    }
-
-    @Test
-    fun closeButton_dismissDialog() {
-        val argument = Bundle()
-        argument.putParcelable(SecurityCodeTooltipDialogFragment.EXTRA_CARD_BRAND, CardBrand.MASTERCARD)
-        var dialog: DialogFragment? = null
-        with(launchFragment<SecurityCodeTooltipDialogFragment>(argument, R.style.OmiseTheme)) {
-            onFragment { fragment ->
-                dialog = fragment
-                assertNotNull(dialog!!.dialog)
-                assertTrue(dialog!!.dialog!!.isShowing)
-            }
-        }
-
-        onView(withId(R.id.close_button)).inRoot(isDialog()).perform(click())
-
-        assertFalse(dialog?.dialog?.isShowing ?: false)
     }
 
     private fun withImageResource(@DrawableRes imageRes: Int): Matcher<View> =
